juce update

+++ MangoChutney/JuceLibraryCode/modules/juce_core/containers/juce_Variant.cpp
@@ -678,12 +678,12 @@ var var::readFromStream (InputStream& input)
 
             case varMarker_Binary:
             {
-                MemoryBlock mb ((size_t) numBytes - 1);
+                MemoryBlock mb (numBytes - 1);
 
                 if (numBytes > 1)
                 {
                     const int numRead = input.read (mb.getData(), numBytes - 1);
-                    mb.setSize ((size_t) numRead);
+                    mb.setSize (numRead);
                 }
 
                 return var (mb);
